dom/xhr/tests/test_xhr_send_readystate.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=814064
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 814064</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=814064">Mozilla Bug 814064</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 814064 **/
var xhr = new XMLHttpRequest();
var firings = 0;
function readyStateHandler() {
  is(xhr.readyState, XMLHttpRequest.OPENED, "Should be in OPENED state");
  ++firings;
}
xhr.onreadystatechange = readyStateHandler;
xhr.open("GET", "");
is(firings, 1, "Should have fired the readystatechange handler");
xhr.send();
is(firings, 1, "Should not have fired the readystatechange handler");
xhr.onreadystatechange = null;
xhr.abort();
is(firings, 1, "Should not have fired the handler no-longer-registered handler");
</script>
</pre>
</body>
</html>
